+++ src/app/(app)/admin/reports/components/ClosedReportsList.tsx
@@ -1,9 +1,7 @@
 "use client"
 
-import Image from "next/image"
 import { useState } from "react"
 import { ArrowRightIcon } from "lucide-react"
-import ReopenRiderReport from "~/components/actions/ReopenRiderReport"
 
 interface Props {
   reports: Array<RiderReport>
@@ -38,11 +36,7 @@ export default function ClosedReportsList({ reports }: Props) {
                 </div>
               </div>
               <div className="relative">
-                {!reportActive ? (
-                  <IdleReportControls open={() => setOpenReport(report)} />
-                ) : (
-                  <ReportActions reportId={report._id} />
-                )}
+                {!reportActive && <IdleReportControls open={() => setOpenReport(report)} />}
               </div>
             </div>
             <div className="w-full">
@@ -81,14 +75,6 @@ const IdleReportContent = ({ report }) => {
           {new Date(parseInt(report.race._id.slice(0, 8), 16) * 1000).toLocaleString()}
         </div>
       </div>
-    </div>
-  )
-}
-
-const ReportActions = ({ reportId }) => {
-  return (
-    <div className="absolute right-0 flex w-fit flex-col justify-center gap-2 align-middle">
-      <ReopenRiderReport reportId={reportId} hackit={true} />
     </div>
   )
 }
