+++ src/app/(app)/appeal/page.tsx
@@ -1,5 +1,5 @@
 import { GetRider } from "~/api"
-import getAuthUser from "~/api/getAuthUser"
+import GetAuthUser from "~/api"
 import PageHeader from "~/components/PageHeader"
 import Pill from "~/components/pills/Pill"
 import BanAppealForm from "./components/BanAppealForm"
@@ -10,7 +10,7 @@ export const metadata = {
 }
 
 export default async function Page() {
-  const user = await getAuthUser()
+  const user = await GetAuthUser()
   const rider = await GetRider(user.guid)
 
   return (
