+++ src/app/(app)/leagues/[leagueId]/signup/page.tsx
@@ -1,5 +1,5 @@
 import { GetConstantOEMBikes, GetConstantServers, GetLeague } from "~/api"
-import getAuthUser from "~/api/getAuthUser"
+import GetAuthUser from "~/api"
 import PageHeader from "~/components/PageHeader"
 import LeagueSignupForm from "../../_components/LeagueSignupForm"
 
@@ -9,7 +9,7 @@ export const metadata = {
 }
 
 export default async function Page({ params: { leagueId } }) {
-  const user = await getAuthUser()
+  const user = await GetAuthUser()
   const league = await GetLeague(leagueId, user.token)
   const oems = await GetConstantOEMBikes(user.token)
   const servers = await GetConstantServers(user.token)
